+++ src/configs/tsdoc.ts
@@ -1,11 +1,11 @@
 import tsdocPlugin from 'eslint-plugin-tsdoc';
 
-import type { Config, ESLint } from '../core';
+import type { Config, Plugin } from '../core';
 
 export const tsdocRecommended: Config = {
   name: 'poupe/tsdoc-recommended',
   plugins: {
-    tsdoc: tsdocPlugin as unknown as ESLint.Plugin,
+    tsdoc: tsdocPlugin as unknown as Plugin,
   },
   rules: {
     'tsdoc/syntax': 'error',